fix: viewPage works
