+++ src/core_modules/capture-core/reducers/descriptions/currentSelections.reducerDescriptionGetter.js
@@ -11,9 +11,6 @@ import {
     actionTypes as viewEventActionTypes,
 } from '../../components/Pages/ViewEvent/viewEvent.actions';
 import {
-    actionTypes as viewEventPageSelectorActionTypes,
-} from '../../components/Pages/ViewEvent/ViewEventSelector/ViewEventSelector.actions';
-import {
     actionTypes as crossPageActionTypes,
 } from '../../components/Pages/actions/crossPage.actions';
 import {
@@ -226,25 +223,61 @@ export const getCurrentSelectionsReducerDesc = (appUpdaters: Updaters) => create
     //     categories: undefined,
     //     categoriesMeta: undefined,
     // }),
-    [viewEventPageSelectorActionTypes.RESET_ORG_UNIT_ID]: (state) => {
-        const orgUnitId = null;
-        const newState = { ...state, orgUnitId };
-        newState.complete = false;
-        return newState;
-    },
-    [viewEventPageSelectorActionTypes.RESET_CATEGORY_OPTION]: (state, action) => {
-        const { categoryId } = action.payload;
-        return resetCategoryOption(state, categoryId);
-    },
-    [viewEventPageSelectorActionTypes.RESET_ALL_CATEGORY_OPTIONS]: state => ({
-        ...state,
-        categories: undefined,
-        categoriesMeta: undefined,
-    }),
-    [newEventSelectorActionTypes.RESET_ORG_UNIT_ID]: (state) => {
-        const orgUnitId = null;
-        const newState = { ...state, orgUnitId };
-        newState.complete = false;
+    // [viewEventPageSelectorActionTypes.RESET_ORG_UNIT_ID]: (state) => {
+    //     const orgUnitId = null;
+    //     const newState = { ...state, orgUnitId };
+    //     newState.complete = false;
+    //     return newState;
+    // },
+    // [viewEventPageSelectorActionTypes.RESET_CATEGORY_OPTION]: (state, action) => {
+    //     const { categoryId } = action.payload;
+    //     return resetCategoryOption(state, categoryId);
+    // },
+    // [viewEventPageSelectorActionTypes.RESET_ALL_CATEGORY_OPTIONS]: state => ({
+    //     ...state,
+    //     categories: undefined,
+    //     categoriesMeta: undefined,
+    // }),
+    // [lockedSelectorActionTypes.SET_EMPTY_ORG_UNIT_BASED_ON_URL]: (state) => {
+    //     const newState = {
+    //         ...state,
+    //         orgUnitId: null,
+    //     };
+    //     return newState;
+    // },
+    // [lockedSelectorActionTypes.SET_ORG_UNIT]: setOrgUnit,
+    // [lockedSelectorActionTypes.SET_PROGRAM_ID]: (state, action) => {
+    //     const programId = action.payload;
+    //     const newState = { ...state, programId };
+    //     newState.complete = false;
+    //     return newState;
+    // },
+    // [lockedSelectorActionTypes.SET_CATEGORY_OPTION]: (state, action) => {
+    //     const { categoryId, categoryOption } = action.payload;
+    //     return setCategoryOption(state, categoryId, categoryOption);
+    // },
+    // [lockedSelectorActionTypes.RESET_CATEGORY_OPTION]: (state, action) => {
+    //     const { categoryId } = action.payload;
+    //     return resetCategoryOption(state, categoryId);
+    // },
+    // [lockedSelectorActionTypes.RESET_ALL_CATEGORY_OPTIONS]: state => ({
+    //     ...state,
+    //     categories: undefined,
+    //     categoriesMeta: undefined,
+    // }),
+    // [lockedSelectorActionTypes.RESET_ORG_UNIT_ID]: (state) => {
+    //     const orgUnitId = null;
+    //     const newState = { ...state, orgUnitId };
+    //     newState.complete = false;
+    //     return newState;
+    // },
+
+
+    [lockedSelectorActionTypes.SET_ORG_UNIT]: (state, action) => {
+        const newState = {
+            ...state,
+            orgUnitId: action.payload.id,
+        };
         return newState;
     },
     [newEventSelectorActionTypes.SET_ORG_UNIT]: setOrgUnit,
